Create a skeleton for navigation between pages
A `ViewModel` base is created upon which all other view models can be based.

The `ShellViewModel` handles logic for navigating between pages. Each page has
an empty view model with nothing but a content string.

A string helper class is created for doing basic operations on strings such
as titleization.

The folder structure is changed, putting Core into a separate folder next to its
tests.

The build matrix is also expanded to include builds for both x86 and ARM CPUs.