Move to Portable Library instead of Universal Library
This should, in the future, make it possible to build on Linux and OS X via
Mono. However, it does not currently build as the current version of Mono does
not support PCL v5.0. I did try to build on PCL v4.5 by targeting .NET 4.5
instead of 4.6 but that got me into issues with the Prism Nuget package.

This also adds a unit test for the shell view model.

It also changes the Slack channel where build status is reported from #stoffi
to #general, since that is the default channel and it's a hassle to always have
to change. Other projects should also change the channel to #general to make
it more convenient.