Rebase on Template 10
By rebasing on Template 10 most boilerplate code can be removed. This
means that the navigation code is removed from the project and most
viewmodels are thus removed from Core.

There may be some code that is currently in the UWP app that can be
moved back into Core, though. But that is an exercise for another day.